+++ src/gnosis-safe/utils/safe-provider.js
